Check for grafana to be ready before reseting the admin password
The grafana reset admin password tries to migrate the grafana db
before applying any change. On slow systems, when monit detects
that grafana is ready (via pid), it runs the post-start, but
might happen that the grafana process is still migrating the db,
so the reset admin password will fail due to a conflict.

Grafana does not start listening until the migration is done, so here
we are checking for the grafana port to be listening before running
the reset admin password. Using this, we ensure that there is no
conflict between the main grafana process and the reset admin
command.