Allow to set a unique subscription id for the firehose-exporter
This commit adds a new parameter called
`firehose_exporter.doppler.use_unique_subscription_id`. If it's true then the
BOSH VM id will be added as a suffix to the subscription id.

The default value of `firehose_exporter.doppler.subscription_id` will be set to
`prometheus` as if this is not set the firehose-exporter will still use that as
a default value. This makes it clearer to use it together with the
`use_unique_subscription_id` flag.