Allow cert contents in prometheus scrape config tls config
Prometheus wants tls config to be file paths, but it is difficult
directly put certificates on a vm while deploying with bosh.

If cert_contents are provided, the pre-start script will write the
contents into cert files that prometheus can use when it starts

Prometheus validates that there are not extra keys in its configuration,
so the extra cert contents keys need to be removed.

[#169551239](https://www.pivotaltracker.com/story/show/169551239)

Co-authored-by: Conor Nosal <cnosal@pivotal.io>