Enable Grafana plugins installation
Grafana plugins can be installed in the UI only if the plugin
directory (defined in grafana.ini) is writable for vcap user.
This commit moves that directory to the persistent disk (so that
plugins are not lost when the instance is recreated) and links
all existing plugins from grafana_plugins package to that directory.